#9: rename folder name using camel case
